Fix a couple of things for deployment and remove some unused components;
