+++ fabioalmer/package.json
@@ -47,6 +47,7 @@
     "react-phone-number-input": "^3.3.7",
     "react-select": "^5.7.7",
     "react-tooltip": "^5.25.2",
+    "sharp": "^0.33.2",
     "zod": "^3.22.4"
   },
   "devDependencies": {
